requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.2.0
pydantic>=2.3.0
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.1
pytest-asyncio>=0.21.1
beanie>=1.20.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6